crypto: move crypto::Level to crypto/mod.rs
