+++ quiche/src/crypto/mod.rs
@@ -36,3 +36,26 @@ mod boringssl_openssl;
     feature = "openssl"
 ))]
 pub use boringssl_openssl::*;
+
+use crate::packet;
+
+#[repr(C)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum Level {
+    Initial   = 0,
+    ZeroRTT   = 1,
+    Handshake = 2,
+    OneRTT    = 3,
+}
+
+impl Level {
+    pub fn from_epoch(e: packet::Epoch) -> Level {
+        match e {
+            packet::Epoch::Initial => Level::Initial,
+
+            packet::Epoch::Handshake => Level::Handshake,
+
+            packet::Epoch::Application => Level::OneRTT,
+        }
+    }
+}
